pyproject.toml
[project]
name = "browser-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "browser-use>=0.12.2",
    "google-api-python-client>=2.188.0",
    "google-auth>=2.48.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "gspread>=6.2.1",
    "jinja2>=3.1.6",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.9",
    "langfuse>=2.54.0",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pandas>=2.3.3",
    "openpyxl>=3.1.5",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "pydantic-settings>=2.13.1",
    "arq>=0.27.0",
    "redis>=5.3.1",
    "aioboto3>=15.5.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
]

[tool.pyright]
include = ["src"]
extraPaths = ["."]
# Helps Pylance/Pyright recognize absolute imports starting with 'src.'
# without incorrectly inferring 'src' as the root.
